feat: restructure Settings into sub-pages, add stream state machine
- Settings root is now a nav list + 10 typed sub-pages (Servers, Appearance,
  Cameras View, Streaming, Events, Notifications, Downloads, Advanced,
  Device Capabilities, About), each with a collapsing top app bar. Servers
  gained a per-row connectivity Test button; Advanced gained a live
  CPU/GPU/RAM/Uptime system-stats view (replacing the raw JSON dump) with
  an auto-refresh option.
- Live camera tiles (WebRTC/RTSP/Snapshot) now share an explicit
  LiveStreamState state machine + StreamOverlay (poster/spinner/error/retry)
  instead of ad-hoc isLoading/statusText tracking, with auto-reconnect and
  a "show last image while loading" setting.
- Fixed: RTSP auth using the wrong host, WebRtcLiveTile missing a fullscreen
  BackHandler, Events camera filter sort order, a duplicated-port RTSP URL
  bug, a bogus VOD "Sub" toggle that always 404'd, and Theme.kt silently
  ignoring the wallpaper-color/palette-style/contrast settings (restored
  from the last known-good committed version — confirmed on-device).
- Events screen now has a real retry-on-failure UI; server password field
  has a show/hide toggle. Both verified on-device against the real server.
- Cleaned up an Android Studio inspection pass: removed a dead LivePlayer.kt
  file and unused EventPill variants, fixed a real UnstableApi lint error,
  a misleading Gradle kotlin{} block nesting, a dead SDK-version check, and
  two unused version-catalog entries.
- Known debt, not fixed here: TrustConfig's "Allow untrusted certs" toggle
  still implements a trust-all X509TrustManager, which conflicts with this
  repo's own rules; no pinned-cert UI exists yet to replace it. See
  memory/project_state.md for details and rationale.

+++ app/src/main/java/net/triton/frigateviewer/feature/settings/AppearanceSheets.kt
@@ -335,7 +335,7 @@ private fun ColorDot(
             Modifier
                 .size(36.dp)
                 .clip(CircleShape)
-                .background(Color(color))
+                .background(Color(color.toInt()))
                 .then(
                     if (selected) {
                         Modifier.border(3.dp, MaterialTheme.colorScheme.onSurface, CircleShape)
@@ -399,7 +399,7 @@ fun AddCustomColorSheet(
                         Modifier
                             .size(48.dp)
                             .clip(CircleShape)
-                            .background(Color(parsedColor)),
+                            .background(Color(parsedColor.toInt())),
                     )
                 }
             }
@@ -419,7 +419,7 @@ fun AddCustomColorSheet(
                         Modifier
                             .size(32.dp)
                             .clip(CircleShape)
-                            .background(Color(color))
+                            .background(Color(color.toInt()))
                             .clickable {
                                 onAdd(color)
                                 onDismiss()
